+++ test/intialize.test.cc
@@ -10,10 +10,10 @@ TEST_CASE("Intializing Tensors", "[int]") {
   auto tensor_1 = Tensor<int32_t, 4>({1, 2, 3, 4});
 
   // Initialize values
-  for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-    for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-      for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-        for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l)
+  for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+    for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+      for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+        for (size_t l = 1; l <= tensor_1.dimension(4); ++l)
           tensor_1(i, j, k, l) = 1000 * i + 100 * j + 10 * k + l;
 
   /*     --------------------------   */
@@ -35,50 +35,50 @@ TEST_CASE("Intializing Tensors", "[int]") {
   }
 
   SECTION("Initializing Values") {
-    for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-        for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-          for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l)
+    for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+        for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+          for (size_t l = 1; l <= tensor_1.dimension(4); ++l)
             REQUIRE(tensor_1(i, j, k, l) == (int)(1000 * i + 100 * j + 10 * k + l));
   }
 
   SECTION("Const Reference Cast") {
     auto const& c_tensor = tensor_1;
-    for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-        for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-          for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l)
+    for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+        for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+          for (size_t l = 1; l <= tensor_1.dimension(4); ++l)
             REQUIRE(c_tensor(i, j, k, l) == (int)(1000 * i + 100 * j + 10 * k + l));
   }
 
   SECTION("Copy Constructor") {
     auto copy_tensor = tensor_1;
-    for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-        for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-          for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l)
+    for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+        for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+          for (size_t l = 1; l <= tensor_1.dimension(4); ++l)
             REQUIRE(copy_tensor(i, j, k, l) == (int)(1000 * i + 100 * j + 10 * k + l));
   }
 
   SECTION("Move Constructor") {
     auto move_tensor = std::move(tensor_1);
-    for (uint32_t i = 1; i <= tensor_1.dimension(1); ++i)
-      for (uint32_t j = 1; j <= tensor_1.dimension(2); ++j)
-        for (uint32_t k = 1; k <= tensor_1.dimension(3); ++k)
-          for (uint32_t l = 1; l <= tensor_1.dimension(4); ++l)
+    for (size_t i = 1; i <= tensor_1.dimension(1); ++i)
+      for (size_t j = 1; j <= tensor_1.dimension(2); ++j)
+        for (size_t k = 1; k <= tensor_1.dimension(3); ++k)
+          for (size_t l = 1; l <= tensor_1.dimension(4); ++l)
             REQUIRE(move_tensor(i, j, k, l) == (int)(1000 * i + 100 * j + 10 * k + l));
  }
 
  SECTION("Value Constructor") {
     auto ones = Tensor<int32_t, 3>({2, 3, 4}, 1);
-    for (uint32_t i = 1; i <= ones.dimension(1); ++i)
-      for (uint32_t j = 1; j <= ones.dimension(2); ++j)
-        for (uint32_t k = 1; k <= ones.dimension(3); ++k)
+    for (size_t i = 1; i <= ones.dimension(1); ++i)
+      for (size_t j = 1; j <= ones.dimension(2); ++j)
+        for (size_t k = 1; k <= ones.dimension(3); ++k)
           REQUIRE(ones(i, j, k) == 1);
     auto twos = Tensor<int32_t, 3>({2, 3, 4}, 2.3f);
-    for (uint32_t i = 1; i <= twos.dimension(1); ++i)
-      for (uint32_t j = 1; j <= twos.dimension(2); ++j)
-        for (uint32_t k = 1; k <= twos.dimension(3); ++k)
+    for (size_t i = 1; i <= twos.dimension(1); ++i)
+      for (size_t j = 1; j <= twos.dimension(2); ++j)
+        for (size_t k = 1; k <= twos.dimension(3); ++k)
           REQUIRE(twos(i, j, k) == 2);
  }
 
@@ -87,10 +87,10 @@ TEST_CASE("Intializing Tensors", "[int]") {
     std::deque<int32_t> container{};
     for (int i = 0; i < 24; ++i) container.push_back(i);
     Fill(naturals, container);
-    for (uint32_t i = 1; i <= naturals.dimension(1); ++i)
-      for (uint32_t j = 1; j <= naturals.dimension(2); ++j)
-        for (uint32_t k = 1; k <= naturals.dimension(3); ++k)
-          REQUIRE((uint32_t)naturals(i, j, k) == (i - 1) * 12 + (j - 1) * 4 + k - 1);
+    for (size_t i = 1; i <= naturals.dimension(1); ++i)
+      for (size_t j = 1; j <= naturals.dimension(2); ++j)
+        for (size_t k = 1; k <= naturals.dimension(3); ++k)
+          REQUIRE((size_t)naturals(i, j, k) == (i - 1) * 12 + (j - 1) * 4 + k - 1);
  }
 }
